Implement update by name in the .update function
